# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Outdoor Event Hub API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /events/{id}/sponsors:
    get:
      operationId: listEventSponsors
      tags: [sponsors]
      summary: List sponsors for an event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Sponsors
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Sponsor"
    post:
      operationId: addEventSponsor
      tags: [sponsors]
      summary: Assign a sponsor to an event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventSponsorInput"
      responses:
        "201":
          description: Assigned
  /events/{id}/sponsors/{sponsorId}:
    delete:
      operationId: removeEventSponsor
      tags: [sponsors]
      summary: Remove a sponsor from an event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: sponsorId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Removed
...
  /sponsors:
    get:
      operationId: listSponsors
      tags: [sponsors]
      summary: List all sponsors with stats
      responses:
        "200":
          description: Sponsors with analytics
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SponsorWithStats"
    post:
      operationId: createSponsor
      tags: [sponsors]
      summary: Create a sponsor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SponsorInput"
      responses:
        "201":
          description: Created sponsor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sponsor"
  /sponsors/{id}/report/export:
    get:
      operationId: exportSponsorReport
      tags: [sponsors]
      summary: Download sponsor impression report as CSV
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: CSV file
          content:
            text/csv:
              schema:
                type: string
        "404":
          description: Sponsor not found
...
  /sponsors/{id}:
    get:
      operationId: getSponsor
      tags: [sponsors]
      summary: Get a sponsor by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Sponsor with stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SponsorWithStats"
        "404":
          description: Not found
    patch:
      operationId: updateSponsor
      tags: [sponsors]
      summary: Update a sponsor
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SponsorInput"
      responses:
        "200":
          description: Updated sponsor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sponsor"
    delete:
      operationId: deleteSponsor
      tags: [sponsors]
      summary: Delete a sponsor
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Sponsor:
      type: object
      required: [id, name, type, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
          enum: [cafe, restaurant, camping, hotel, gym, shop, other]
        website:
          type: ["string", "null"]
        instagram:
          type: ["string", "null"]
        facebook:
          type: ["string", "null"]
        logoUrl:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        discountCode:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    SponsorInput:
      type: object
      required: [name, type]
      properties:
        name:
          type: string
          minLength: 1
        type:
          type: string
          enum: [cafe, restaurant, camping, hotel, gym, shop, other]
        website:
          type: string
        instagram:
          type: string
        facebook:
          type: string
        logoUrl:
          type: string
        description:
          type: string
        discountCode:
          type: string
    SponsorWithStats:
      type: object
      required: [id, name, type, createdAt, pageViews, registrationsFromPage, approvedRegistrations, conversionRate, eventsCount]
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
          enum: [cafe, restaurant, camping, hotel, gym, shop, other]
        website:
          type: ["string", "null"]
        instagram:
          type: ["string", "null"]
        facebook:
          type: ["string", "null"]
        logoUrl:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        discountCode:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        pageViews:
          type: integer
        registrationsFromPage:
          type: integer
        approvedRegistrations:
          type: integer
        conversionRate:
          type: number
        photoImpressions:
          type: integer
        eventsCount:
          type: integer
...
    EventSponsorInput:
      type: object
      required: [sponsorId]
      properties:
        sponsorId:
          type: integer
...